requirements.txt
Django~=3.2.5
SpeechRecognition==3.8.1
pyaudio~=0.2.13
picovoice
pvrecorder
pyyaml
